requirements.txt
# Core dependencies
streamlit==1.31.0

# Google Gemini
google-generativeai==0.3.2

# ElevenLabs for text-to-speech
elevenlabs==1.0.0

# Firecrawl for web scraping
firecrawl-py==0.0.16
